Move MediaInfo struct to http package
